requirements.txt
asgiref==3.8.1
attrs==24.2.0
Django==5.1.2
flake8==5.0.4
flake8-annotations==2.9.1
flake8-quotes==3.3.1
flake8-variables-names==0.0.5
mccabe==0.7.0
pep8-naming==0.13.2
pycodestyle==2.9.1
pyflakes==2.5.0
sqlparse==0.5.1
requests~=2.32.3
environ~=1.0
python-dotenv~=1.0.1